Do not export raw metrics if reading failed
Most of the metrics are never exported if they cannot be read. This
extends this behaviour to the raw sensor values. This is useful for
situation where ipmitool fails, e.g. on most virtual machines. For
example the raw values would indicate a power supply failure, when
really this is not the case.